refactor: reorganize md2img module and update import paths
